DSS-1480 : Integration OpenPDF DSS-1444 : PDFDocumentValidator does not report when it could not parse the provided PDF

+++ dss-pades-openpdf/src/main/java/eu/europa/esig/dss/pdf/openpdf/visible/ImageOnlySignatureDrawer.java
@@ -46,8 +46,8 @@ public class ImageOnlySignatureDrawer extends AbstractITextSignatureDrawer {
 		if (height == 0) {
 			height = (int) image.getHeight();
 		}
-		width *= (zoom / 100);
-		height *= (zoom / 100);
+		width *= ((float) zoom / 100);
+		height *= ((float) zoom / 100);
 
 		if (Utils.isStringNotBlank(signatureFieldId)) {
 			appearance.setVisibleSignature(signatureFieldId);
